+++ packages/fighting-design/header/index.ts
@@ -4,7 +4,9 @@ import { install } from '../_utils'
 
 export const FHeader = install(Header)
 
-export type FHeaderInstance = InstanceType<typeof Header>
+export type HeaderInstance = InstanceType<typeof Header>
+
+export type { HeaderPropsType } from './src/header'
 
 declare module 'vue' {
   export interface GlobalComponents {